lambdas/calendar_fetcher/requirements.txt
boto3==1.34.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0
pytz==2023.3
requests==2.31.0
